requirements.txt
asgiref==3.2.10
Django==3.0.1
djangorestframework==3.10.1
Pillow==7.2.0
pytz==2019.1
sqlparse==0.3.0
